+++ src/app/models/press.models.ts
@@ -1,0 +1,11 @@
+export interface Press {
+	_id: string;
+	nr: string;
+	desc: string;
+	source: string;
+	date: Date;
+	author: string;
+	fileExtension: string;
+	link: string;
+	quote: string;
+}
