+++ src/app/models/sponsors.models.ts
@@ -1,0 +1,11 @@
+export interface Event {
+	event: string;
+	share: number;
+}
+
+export interface Sponsor {
+	events: Event[];
+	image: string;
+	name: string;
+	url: string;
+}
